requirements.txt
altgraph                  0.17.3    
asgiref                   3.7.2     
backports.zoneinfo        0.2.1     
certifi                   2022.12.7 
charset-normalizer        2.1.1     
colorama                  0.4.6     
comtypes                  1.2.0     
contourpy                 1.1.1
cycler                    0.12.1
darkdetect                0.8.0
distlib                   0.3.7
exceptiongroup            1.2.0
filelock                  3.12.4
fonttools                 4.47.0
future                    0.18.2
gui2py                    0.9.3-py2x
idna                      3.4
imageio                   2.34.0
importlib-metadata        7.0.1
importlib-resources       6.1.1
imutils                   0.5.4
iniconfig                 2.0.0
keyboard                  0.13.5
kiwisolver                1.4.5
lazy_loader               0.3
matplotlib                3.7.4
numpy                     1.23.4
opencv-contrib-python     4.8.1.78
opencv-python             4.8.1.78
packaging                 23.2
pefile                    2022.5.30
pillow                    10.2.0
pip                       22.3
platformdirs              3.11.0
pluggy                    1.3.0
psutil                    5.9.8
PyAutoGUI                 0.9.53
PyGetWindow               0.0.9
pygrabber                 0.1
pyinstaller               6.4.0
pyinstaller-hooks-contrib 2024.1
pylibdmtx                 0.1.10
PyMuPDF                   1.23.26
PyMuPDFb                  1.23.22
pynput                    1.7.6
pyodbc                    5.0.1
pyparsing                 3.1.1
PyQt-Fluent-Widgets       1.1.8
PyQt5                     5.15.10
PyQt5-Frameless-Window    0.3.3
PyQt5-Qt5                 5.15.2
PyQt5-sip                 12.13.0
PyRect                    0.2.0
PyScreeze                 0.1.30
pyserial                  3.5
pytesseract               0.3.10
pytest                    7.4.4
python-dateutil           2.8.2
pywin32                   306
pywin32-ctypes            0.2.2
pywinauto                 0.6.8
pyzbar                    0.1.9
requests                  2.28.1
scikit-image              0.21.0
scipy                     1.10.1
setuptools                65.5.0
six                       1.16.0
sqlparse                  0.4.4
tifffile                  2023.7.10
typing_extensions         4.8.0
tzdata                    2023.3
urllib3                   1.26.14
virtualenv                20.24.5
wheel                     0.37.1
wxPython                  4.2.0
zipp                      3.1*7.0
zxing                     1.0-+